requirements.txt
fastapi == 0.115.3
uvicorn == 0.32.0
python-dotenv ~= 1.0.1
httpx ~= 0.27.2
pyjwt[crypto] ~= 2.9.0
pydantic[email] ~= 2.9.2
psycopg == 3.2.3
pandas ~= 2.2.3
openpyxl ~= 3.1.5
scikit-learn ~= 1.5.2
azure-storage-blob ~= 12.23.0
aiohttp ~= 3.10.10
pillow ~= 11.0.0
beautifulsoup4 ~= 4.12.3
lxml ~= 5.3.0
openai == 1.52.2
